style: move rewrite algorithm to separated file
