+++ src/index.ts
@@ -1,12 +1,13 @@
-import kleur from 'kleur'
-import minimist from 'minimist'
 import fs from 'node:fs'
 import path from 'node:path'
 import url from 'node:url'
+import kleur from 'kleur'
+import minimist from 'minimist'
 import prompts from 'prompts'
 import { simpleGit } from 'simple-git'
-import { clearDir, copy, isEmptyDir, OVERRIDE_FILE } from './fs'
+import { clearDir, copy, isEmptyDir } from './fs'
 import { PackageManager, getPackageManager, isValidPackageManagerName } from './package'
+import { rewrite } from './rewrite'
 import { isValidTemplateName, templates, type Template } from './template'
 import { isValidPackageName, isValidProjectName } from './validate'
 
@@ -135,91 +136,11 @@ const cli = async () => {
       baseDir: root,
     })
 
-    const userName = (await git.getConfig('user.name')).value ?? ''
-    const userEmail = (await git.getConfig('user.email')).value ?? ''
-    const projectRepo = userName && userEmail ? `https://github.com/${userName}/${packageName}` : ''
-
-    // read package.json file content to get infos and do some edits
-    const pkg = JSON.parse(fs.readFileSync(path.resolve(root, 'package.json'), 'utf-8'))
-
-    // cache some fields of package.json
-    const _packageName = pkg.name
-    const _userName = pkg.author.name
-    const _userEmail = pkg.author.email
-    const _projectRepo = `https://github.com/${_userName}/${_packageName}`
-    const _projectDesc = pkg.description
-    const _projectKeywords = pkg.keywords.join(',')
-
-    // overwrite some fields of package.json
-    pkg.name = packageName
-    pkg.version = '0.0.0'
-    pkg.description = ''
-    pkg.keywords = []
-    pkg.repository.url = `git+${projectRepo}.git`
-    pkg.homepage = `${projectRepo}#readme`
-    pkg.bugs.url = `${projectRepo}/issues`
-    pkg.bugs.email = userEmail
-    pkg.author.name = userName
-    pkg.author.email = userEmail
-    pkg.author.url = `https://${userName}.github.io/`
-    pkg.contributors = [userName]
-
-    Object.entries(OVERRIDE_FILE).forEach(([key, files]) => {
-      let source: string, target: string
-      switch (key) {
-        case 'packageName':
-          source = _packageName
-          target = packageName
-          break
-        case 'userName':
-          source = _userName
-          target = userName
-          break
-        case 'userEmail':
-          source = _userEmail
-          target = userEmail
-          break
-        case 'repository':
-          source = _projectRepo
-          target = projectRepo
-          break
-        case 'description':
-          source = _projectDesc
-          target = ''
-          break
-        case 'keywords':
-          source = _projectKeywords
-          target = ''
-          break
-        default:
-          throw new Error(`${kleur.red('✖')} Unhandled key "${key}" in OVERRIDE_FILE.`)
-      }
-
-      // travel each file that need to update package name
-      for (const file of files) {
-        // read file content
-        let content = fs.readFileSync(path.resolve(root, file), 'utf-8')
-
-        // overwrite the project name
-        content = content.replaceAll(source, target)
-
-        // write file content
-        fs.writeFileSync(path.resolve(root, file), content)
-      }
-    })
-
     // init project git config
     await git.init()
 
-    // override package.json file content
-    fs.writeFileSync(path.resolve(root, 'package.json'), JSON.stringify(pkg, null, 2))
-
-    // override .all-contributorsrc file content
-    const acs = JSON.parse(fs.readFileSync(path.resolve(root, '.all-contributorsrc'), 'utf-8'))
-    acs.projectName = packageName
-    acs.projectOwner = userName
-    acs.contributors = []
-    fs.writeFileSync(path.resolve(root, '.all-contributorsrc'), JSON.stringify(acs, null, 2))
+    // rewrite some project files
+    await rewrite(root, packageName, git)
 
     // print prompt message
     logger.log()
